parallel version with improved registration
It runs in parallel thanks to Abraham.

It uses the new initial estimates based on orthogonal matrices - so for the embryo I used it only had to run 8 initial
estimates to find the best - and so I set maxItr = 8; -- but not sure what it should be for other embryos.

I also put the matching and the lineage construction code in this program and I save this so it can be used by the MakeLineage code. If I don't find the registration in the 8 init estimates - then I remove the 'poor' matches and try again. For my embryo, this
fixed everything and the registration sigma was less than 10 for the first 100 frames. I put the visualizations there (and the code: